Polish multi-tenant runtime for review and maintainability
- reduce review overhead around the tenancy implementation
- improve code comments for tenant-owned metadata, reload safety, and logical isolation
- clarify management-auth trade-offs and what remains global vs tenant-isolated
- tighten README and sample config wording to match the implemented design
- keep runtime behaviour stable while making the branch easier to review and demo
